Add Windows CI
This adds an MSCV2019 Release build to the
build and test job. The job uses conan
for dependencies other than qt.

We might want to consider also moving to
conan for the linux build at some later
point in time.

Note, Ninja is needed to support ccache.
Also note, ccache does not support "-Zi",
so there is no relwithdebinfo build.

Test: Run CI