Fix capture window flickering when loading a new font (#4420)
In this PR we are fixing the flickering that appeared when
QT-TextRenderer was introduced and that was noticeable at the
beginning of a session.

After an investigation, turned to be that the cause was that when QT
loaded a new font, it messes the openGL state and nothing else was being
drawn in that frame.

So, in this PR:
 - We are calling CleanupGlState() & PrepareGlState() in between
   QT-TextRendering and native GL-ShadeRendering (for each layer).
 - Including these lines in PrepareGlState():
    - glBindTexture(GL_TEXTURE_2D, 0); -> This line fixed the issue.
    - glUseProgram(0);
    They were called before at the beginning anyways, but not in between
    QT and native GL rendering and they should be part of
    PrepareGLState().

Screenshot of the flickering: http://screenshot/7StjuKsDojuyS7j. Nothing
after kZValueTrackText was rendered in native GL.

Test: Loaded a capture, everything works. Take a capture, no flickering
either.