Remove GLAD OpenGL loader (#4456)
The GLAD OpenGL loader used to exist because firstly it was supported by
freetype-gl and secondly it allowed us to chain itself to the Qt OpenGL
loader. So it acted as some kind of bridging code between freetype-gl
and Qt (and was also used by code in OrbitGL).

Since freetype-gl are dependencies anymore we can transition to using
the Qt OpenGL loader everywhere in the code base. This requires all
OrbitGl classes that call OpenGL function to subclass from
QOpenGLFunctions which implements all the OpenGL functions as member
functions.