Unicode support in QtTextRenderer. (#4464)
We now convert the char* we get to QString directly. This gives us
proper unicode aware handling of string length, substrings and splitting
on newlines.

I did not touch the external interface. We still use char* as input.
It might be a good idea to swich that toi QString in a follow up PR.