Add include-what-you-use action
With this action we are applying IWYU to the codebase and create
a PR with the fixes.

This action will be run on every commit on main touch code in src/.
The generated fixes are going to be applied and pushed to a
branch hotfix/iwyu-COMMIT_SHA, where the COMMIT_SHA is the commit
the action run on. Afterwards, a PR is created to merge the
change into main.

Note, that in order to pass the CLA check, we are using the
orbitprofiler-bot service account with its personal access
token.

Test: Add the changes to fork and create a change with a
      violation.