Improve command help box by using unicode characters (#4458)
In this PR we are improving the display box that shows capture window
controls, now that we can use unicode characters.

The appearance was discussed with our UXD and can be seen in
http://screenshot/3ssv786c9jTtPv4. There were a few changes in the
displayed text and the opacity of the box changed from 90.2% to 95.3%
(half transparent than before). We also fixed an scaling issue. Now
text will also scale with the rest of the capture window:
http://screenshot/7CAvc8mXCLmGAUx.

We also took advantage of the local time_graph_layout_ member in the
CaptureWindow and no use it time_graph_->GetLayout() anymore.

Bug: http://b/241923528 & http://b/180312795
Test: Scale the UI, see the box scaled accordingly.