Don't use screen coordinates when drawing
This PR is a fixup of https://github.com/google/orbit/pull/4383.
In that PR we fixed samples drawing in screen coordinates but just for
selected ones (by mistake), so this PR is completing the work.

I had it as draft 3 days ago but I wasn't sure if they were working
correctly as there are still some empty spaces in the callstack sample
bar (http://screenshot/mxsPeiYPopLJvjU). After printing all the values
where the Callstack is drawing lines, all of them are there, so I think
it's an issue related to how OpenGl draw lines.

Anyway, it's not a visual regression so I will merge it.

Test: Load 2 captures, check no visual regression, print all the
returned values from GetBoxPosAndSize, all the expected values are
there.