Fix IWYY-introduce Windows build failure
This also cleans up the Logging.cpp file. So far we were using the
TEMP_FAILURE_RETRY macro which only exists on Linux and polyfilling it
for Windows.

But the usage of this macro is not needed because signal interruptions
are already handled in the libc IO functions (fopen, fclose, fread,
fwrite, etc.)

It would only be needed if we were using Linux syscall wrappers like
(open, close read, write, etc.)

So this remove the usage of the retry logic and also moves the Windows
only debugger output function into a separate file (as it's our usual
policy.)