More Windows build fixes
This commit mainly adds support for MSVC2022 and removes support for MSVC2017.

It also removes support for building Qt from source (which was still
offered as an option in build.ps1).

We can consider automatically installing Qt using aqtinstall in a later
step.

Note, that this is just a first step in fixing the Windows build. More
needs to be done.