Add generic ssh support via flags (#4473)
This is done by adding flags and changing the ConnectToTargetDialog,
which used to work by supplying a Stadia instance and process. The
ConnectToTargetDialog now connects to a remote machine when the ssh
flags are used. Adds a convenience run script for ssh connections and
removes stadia related run scripts. Removes VSI/ConnectToTargetDialog
flags --target_instance and --target_process. Removes the testhub flag
--target_uri and related functionality.

In a next step, there will be a UI for the ssh flags.