Fix autopresets copy error (#4487)
When building a clean build on Windows the build aborted with an error
saying that it can't copy the autopreset to the bin/ subdirectory. The
problem was that already a file with the name `bin` existed.

It turned out this was the also copied DIA SDK DLL. The solution is to
create the directory explicitly.