Add a left margin to TimeGraph
It is assumed at various places that the thread tracks are positioned
at 0 horizontally. To uncover those assumptions, and for easier future
layout modifications, we add a left margin to time graph layout.
It is set to zero, so there is no functional change in this change.
To visualize the effect, the "Left Margin" has been exposed to the
"devmode" layout UI.