Add & integrate ProcessListWidget
This integrates ProcessListWidget into SessionSetupDialog, replacing/
moving functionality away from SessionSetupDialog. Then ProcessManager
is moved from TargetConfiguration.h into Connections.h, which results in
changes to orbitmainwindow, ConnectToLocalWidget and
ConnectToTargetDialog. One of the changes to ConnectToLocalWidget is,
that now a signal is emitted when a new process list is available.