Parse modules for hotpatchable functions. (#4497)
In order to represent functions in Orbit we read information from
various sources: elf files, coff files, pdb's. In the end all of these
produce SymbolInfo protos. This change adds a is_hotpatchable field to
SymbolInfo. Only for elf files we parse for hotpatchable functions and
set the new field to true if appropriate. Msvc also supports
hotpatchable functions but for now we ignore this.

SymbolInfo is translated into FunctionInfo - so this also gets the
is_hotpatchable field.

FunctionInfo is translated to a InstrumentedFunction in the capture
options. The is_hotpatchable field is added there as well.

The rest of the PR is basically adding the new is_hotpatchable field to
all the tests. The only real addition to testing is in ElfFileTest - it
is parsing the hotpatchable function section from a test binary.