Fix ownership in CodeViewer
QApplication::sendEvent doesn't take ownership of the event because it
doesn't need to (it's processed synchronously).

So the call to `unique_ptr::release` leads to a memory leak.

Solution: event on the stack and passed as a simple pointer to the stack
value.

This problem was found using Address and UB Sanitizer™.