Fix UB and a segmentation fault in HttpDownloadManagerTest
This replaces manual memory management for the `HttpDownloadManager` by
a `std::optional`. This avoids that the memory will be leaked (delete
has never been called) and also avoids the need to call the destructor
manually - which can have all kinds of problems.

This has been found using Address and UB Sanitizer™.