Fix usage of absl::bit_cast
`absl::bit_cast` (and the `std::bit_cast`) have to be used "by value".
If they are used to convert pointers, it's not more than a fancy
`reinterpret_cast` and that's what we want to avoid.

This also means `bit_cast` is limited by converting objects of the same
size. It can't read from a char buffer.

So this commit replace the two bit_casts by std::memcpy which is the
non-UB way of reading something from a buffer into a variable that is
trivial.

This has been found by Address and UB Sanitizer™.