Fix segmentation fault in GpuTrackTest (#4517)
The GpuTrackTest instantiates a `GpuTrack` and has to provide a pointer
to `OrbitApp` due to our legacy code structure. Since it is impossible
to instantiate a `OrbitApp` in a unit test we pass in a nullptr
(assuming the `OrbitApp` instance won't be used).

But that's not entirely true. `GpuTrack` calls `OrbitApp::GetStringManager`
to get string manager instance. This leads to a segmentation fault in
unoptimized builds.

The solution is to have a separate constructor argument for the string
manager.