Fix memory leak in SymbolLocationsDialog (#4513)
`QListWidget::takeItem` removes the item from the list and transfer
ownership of the `QListWidgetItem` to the caller - similar to
`unique_ptr::release`.

But that's not what we want here. We want to remove the item from the
list and delete the object representing the item. According to Qt docs
this is done by calling delete on the object.

This was found using Address und UB Sanitizer™.