Add SaveRangeFromArg GMock Action
This adds a new GMock action which helps when mocking a function that
has an absl::Span as one of the parameters.

If we want to capture this Span argument `testing::SaveArg` is not
suitable because it tries to make a copy of the Span which leads to
lifetime issues.

This action is a solution because it makes a deep copy of a range into
an arbitrary container that can be constructed from a range (can be a
vector or a set for example).