Change TrackManager::GetOrCreate* to std::string_view (#4540)
Some GetOrCreate<Something>Track functions in TrackManager used to take
a const std::string&. Changing that to std::string_view was to intrusive
for the huge mechanical change that switched everything to
std::string_view.

So this PR follows up on the missing parts. It enables transparent
comparison for `async_tracks_` and `variable_tracks` and rearranges the
code so that it follows our return early guideline.