Fix out of bounds access in SetOrbitThreadsInTarget
`SetOrbitThreadsInTarget` calls `SetOrbitThreads` in our user space
instrumentation library through ptrace. To keep things simple
`SetOrbitThreads` takes a fixed number of parameters which represent the
TIDs of the injected threads (mostly from gRPC) to filter out events
from them.

The TIDs of the injected threads are determined by comparing a list of
TIDs from after the injection to a list from before the injection.

The code assumes that the resulting list contains at least 6 TIDs. But
that recently changed. An update to the gRPC library reduced the number
of threads and this now leads to an out of bounds access in the hard
coded list.

This PR solves that by checking the size of the given list and providing
a fall back value of -1 if the number of threads given in the list is
shorter than 6.

Note that this still doesn't solve the case when we injected more than 6
threads. To solve this we probably have to change the data structure in
the instrumentation library (from an array to a vector or similar) and I
am not sure if this affects the performance in any way. That's why this
is only fixing the OOB access.