Disable two clang-tidy rules.
"modernize-deprecated-headers" will wrongly include <cstdint>/<cstddef>
instead of <stdint.h>/<stddef.h> for usages of uint32_t, size_t, etc..
Those headers do not define the c symbols, but the "std::" ones.
So this rule is agains IWYU.

"readability-identifier-length" is a valid rule, however, we do have
quite a number of places where it is not applicable, e.g. the name
"id" is used quite a lot.

Test: Run clang-tidy