Apply "readability-identifier-naming" clang-tidy check
Let's agree on naming: Only constexpr is "kCamelCase",
all the other variables should be snake_case. That
is a bit annoying for "static const" variables, that
are effectively also constants, but clang-tidy does
not support that, yet.

This change does intentionally not touch method names,
yet. There will be likely a follow-up for that.

This needs to be reviewed a bit more carefully. At
some places, I switched to constexpr, when appropriate,
or added a const modifier.

Test: Compile