Add simple SSH server to CI (#4573)
This is adding a docker based SSH server that can be used by tests to
our CI. Connection paramaters are transferred via environment variables
to the tests.

So far the docker container is built each time. This takes a couple of
seconds. Later on we can optimize that by caching them in GitHub's
Docker registry.

I have called it the "simple SSH Server" because I expect that we might
have other SSH server with different behaviour for advanced testing of
the SessionSetup code.