Add tests for ServiceDeployManager
This is adding tests for all 3 deployment methods of
ServiceDeployManager.

Two of them need to upload some artifacts to a remote machine that behave
like a (packaged) OrbitService. These don't need to be the real
artifacts. The SSH server containers that we use for testing are
prepared to work with simple fakes, like shell scripts instead of a
binary OrbitService.