Add promises to StateMachineHelper
This is adding two promises to the StateMachineHelper which will allow
users of StateMachineHelper to hand out futures to their users for
both the starting phase and the stopping phase.

This commit does not contain any usages (this will come in a separate
PR), but just as an example: The orbit_ssh::Session type has a
`ConnectToServer` and a `Disconnect` method. With this change both
methods can return a future that callers can wait for to determine when
a connection has been made / when a connection has been disconnected.