Return futures in Task and Tunnel's public methods
Similar to `orbit_ssh_qt::Session`, this is changing the public methods
ot `Task` and `Tunnel` to return futures as far as possible.

This includes:
- `Task::Start`
- `Task::Stop`
- `Task::Write`
- `Tunnel::Start`
- `Tunnel::Stop`

The `Start`/`Stop` methods are trivial because they rely on the
facilities provided by `StateMachineHelper`. `Task::Write` is a little
more contrived, because it requires `Task` to do some state management.