Add orbit_qt_test_utils::WaitFor (#4596)
`WaitFor` is a simple function that takes a future and blocks until it
completes. In contrast to `future.Get()` (which also blocks until it
completes), `WaitFor` spins up an event loop in the background and
processes Qt events that have been scheduled for the same thread.

This is very helpful to test asynchronous APIs in a synchronous way.