Clang-tidy CI (run #2)
In order to successfully run clang-tidy, we actually need to have
the auto-generated code available. As we don't have a step for,
we move the whole CI step to the build phase.

Let's also switch the action to report clang-tidy warnings,
as this allows us to have full control over clang-tidy.

Test: Tested on fork.