Use the new TemporaryDirectory (#4569)
So far we only had `TemporaryFile` which creates and opens a temporary
file and deletes it when the instance gets out of scope. The intention
is that users of `TemporaryFile` operate on the file descriptor provided
by the type.

But we have many APIs that operate on file names, not on file
descriptors. To "solve" that problem tests for these APIs call
`TemporaryFile::CloseAndRemove` to delete the file and then use the file
name that is provided by `TemporaryFile::file_path`. The problem is that
this file path is not guaranteed to be free after CloseAndRemove has
been called (also pointed out here:
https://en.cppreference.com/w/cpp/io/c/tmpnam)

Not sure how often this problem really occurs but I had test failures
due to that in the past. The solution is to move these tests over to
`TemporaryDirectory` which is a new class. It offer a temporary
directory that is cleaned up when the instance gets out of scope.