Silence clang-tidy in ORBIT_CHECK macro
First this is disabling the avoid-do-while-loop check for macros because
it's a common way to create macro the behaves like a function (expects a
semicolon at the end).

Second this is explicitly ignoring the return value of the `fprintf` to
silence another check about using return values. It also adds `std::` to
this functions and replaces it by the more suitable and faster `fputs`
function.