Add run_ssh_containers.sh helper
This is a helper script that simplifies launching the tests that need a
running SSH server. The script uses containers (Linux only) to spin up
an independent SSH server and also shuts it down when it's done.

The commit also rearranges some files. The directory `ssh_server_simple`
in `contrib/docker/` gets renamed to `ssh_server`. The first intentation
was to have separate directories for each SSH server, but that's
cumbersome because they share a lot of the same files (like host keys,
etc.). So instead the `Dockerfile` gets renamed to `Dockerfile.simple`
such that there can be additional Dockerfiles in the futures. All users
get adjusted as well.