Replace invalid usage of bit_cast in PerfEvent code
`absl::bit_cast` - when used to cast pointers - is no different to
`reinterpret_cast`. There is no added safety because the pointer gets
casted, not the pointee.

For the added safety `bit_cast` needs to be used with values.

In this particular case that doesn't work because `bit_cast` requires
the source type to be of the same size as the target type. So this
doesn't work with memory buffers.

So this change falls back to using memcpy which is AFAIK the only valid
way of doing this.