Fix bug in WaitFor test utility (#4642)
The r-value overload of `GetValue` had a copy-and-paste error. So far it
has not been instantiated, that's why it didn't lead to a compiler
error.

So this commit fixes the bug, and adds a test that instantiates the
overload.

It is also adding another helper function to simplify test code.