Fix introspection crash (#4647)
There is a weird pattern in `CaptureDataHolder` that returns a
reference from a pointer that could be null. It seems there is
already a plan to fix this pattern (b/234095077), but in the
meantime Orbit sometimes crashes when selecting timers in the
introspection window. This PR adds a `ProvideScopeId` method to
`CaptureDataHolder` which checks that `capture_data_` is valid
before dereferencing it and otherwise returns std::nullopt.