Auto unwrap in Executor (#4655)
When first introducing Futures I opted for making unwrapping explicit
and see first how the feature is used. Now, after more than a year, I
don't see any downside anymore to automatic unwrapping.

How it works
------------

An `Executor` has a `Schedule` function that returns a
`Future<T>` with `T` being determined by the return type of the action
that gets scheduled. If `T` is of type `Future<U>` then `Schedule` used
to return a `Future<Future<U>>`. Now with auto unwrapping, it returns a
`Future<U>`.

That means there is no longer a way to observe when this
particular task finishes executing on a particular executing (for
example when it frees up thread in a thread pool) But so far there
is and never has been a use-case for that in the code base.