Allow thread selection alongside time range
This allows users to select a thread either on it's own or in
conjunction with a time range selection. The UI needs some improvements
to make this more obvious, but the functionality is all there in this
PR. It is still hidden behind the flag --time_range_selection so will
nto affect any current workflows.

Bug: http://b/240111360