Add WhenValueOrCanceled
This adds a helper function which takes a `StopToken` and a future and
returns a future that completes when either the StopToken indicates a
requested stop or the future completes.

It's helpful to make futures appear "cancelable" and to write more
readable code with clear intent.

The implementation is basically a copy of `WhenValueOrTimeout`.