Accelerate text rendering.
The algorithm fitting the text into a given width was slow as reported
here: https://github.com/google/orbit/issues/4626

This PR replaces the code with the methods Qt provides for this purpose.

Since Qt does things different from what was implemented before we get
some changed behaviour:
Most notably we don't render text into any timers that are too small to
fit at least "W...". Also, instead of truncating short texts, we now
truncate and add "...".

Since the bahaviour is changed it is a bit hard estimate the performance
gain. Measurements suggest ~4-5x with test cases that previously did a
lot of calls in the code fitting the strings into a given width. I think
this largely solves the issue in the bug above.