Improve font placement.
For unknown reasons the placement of the fonts in the timer intervals is
a bit off under Windows. Compare https://github.com/google/orbit/issues/4627.

This addmittedly is a hack to compensate for these differences. It would
probably be cleaner to either understand the root cause or to subtract a
margin from the timer box and render the text inside that. But this is a
quick fix and it solves the issue.