Use absl::Mutex in TrackManager
TrackManager is a class that is not as easy as it should be to navigate. The main complexity lies in the fact that there are two threads that are accessing the class. The capture thread is inserting timers and therefore creating the tracks when needed and the main thread is reading the tracks to sort then and render the visible ones (a detailed picture can be found in http://go/stadia-orbit-track-creation#heading=h.vpgt3chmpmpb).

We had issues in the class because of multi-threading (and currently it seems Daniel found a new one), so in this PR I'm cleaning this class by replacing the recursive_mutex with an absl::Mutex. I don't see how this could solve the crash found by Daniel but at least the class might be easier to understand when looking for the issue.

We were using a recursive_mutex because some public methods were calling another ones where both of then needs the mutex to be locked. But as we don't have recursion, this can be solved by splitting some methods.

In this PR we are:
- Guarding all the containers that needs to be locked (ABSL_GUARDED_BY)
- Locking all the public methods that are accessing the containers.
- Not locking the private methods, as they are only called by the class itself.
- Manually assuring that the locked methods don't call another public method that lock the mutex.

I'm not sure about if the last point is the best thing to do but I feel that this is an improvement, do you have another suggestion?

Test: Load a capture with all kind of tracks. Also take a capture.