Cache FindModuleByAddress
When creating the different sampling views, there
are a lot of querries to FindModuleByAddress, which is
performing a binary search.

This change saves the results such that the binary search
happens only once per address.

Follow-up: Also do caching for FindFunctionByAddress and
similar methods. Those are actually more promesing
in terms of performance improvements.

Test: Tested on 5-min capture
Bug: http://b/264437402