Cache FindFunctionByAddress (#4687)
When creating the different sampling views, there
are a lot of querries to FindFunctionByAddress, which
is performing a binary search.

This change saves the results such that the binary
search happens only once per address.

Measuring improvements here is kinda hard, as there are
many things that rely together. This change alone
is "only" about a 1.05x speedup. However, combined with
PR #4686, as well as some further caching, we get ~2x.

Test: Tested on 5-min capture
Bug: http://b/264437402