Optimize ModuleManager
This change is actually tackling two things:

1. Following go/fast/55, we replace the node_hash_map, we used
for pointer stability by a flat_hash_map of unique_ptrs.

2. GetModuleByModuleInMemoryAndAbsoluteAddress uses strings
as keys, which's hash computation comes with some cost.
Absolute address was only used for validation.
We now also cache the modules for an absolute address.

This change alone, is about a 1.15x speedup.

Test: 5-min test capture
Bug: http://b/264437402