Deactivate thread bar based on selections
This change makes the thread bars gray when another thread is selected
or when it is outside of the time range.  I also changed
TicksToMicroseconds to make it possible to be negative if the time given
was before the start of the capture.

Screenshot: https://screenshot.googleplex.com/9i3ZXgcDQo9rUjT

Bug: http://b/240111360