Fix crash on load capture. (#4698)
We are sorting the tracks according to the number of timers they
contain. During a load of a capture this number changes.

Previously we were doing a sort on the tracks and a different thread was
changing the number of timers and therefore the order used in the sort.
This does not end well.

The remedy is to take a snapshot of the number of timers such that sort
can operate on a stable order.