Add memory measurements for the Batcher (#4701)
Add information about Batcher memory requirements to the debug output.
This reports the actual memory required by all of our BlockChain structures
(geometry, colors, and picking information) including the empty space in the
blocks.

Currently only the memory required by the TimeGraph batcher is reported
as this is the critical one (it contains all the timers). It's easy to add the UI
batcher in the future if needed.