Fix update requests for thread states mouse over (#4703)
Any changes that require redraw of a CaptureViewElement should call
RequestUpdate(). This was missing in the mouse-over handler of the
ThreadStateBar.

This will currently not change anything about the behavior, as the call to
app_->set_selected_thread_state_slice already forces a complete redraw
of the capture window. In the future, we would like to move to a state where
partial redraws are possible, and thus it's not enough to force this globally.