Fix Executor lifetime handling (#4670)
The `Executor` type has 2 facilities which might call `Schedule` any
moment in time. These are:

- `ScheduleAfter` (member function)
- `TrySchedule` (free function)

Hence `Executor::Schedule` must at the very least be thread-safe, but it can also happen that `ScheduleAfter` or
`TrySchedule` try to call `Executor::Schedule` while the executor's
lifetime has already ended or worse while the executor's destructor is
currently running on a different thread.

We learned about this problem the hard way by a crash found in release
testing and quickly fixed it by requiring any `Executor` to be a shared
pointer. `ScheduleAfter` and `TrySchedule` would keep a weak pointer of
the executor and only call `Schedule` if the executor was still alive.

But this was a hack and is not great type design because this shared
pointer was part of the public API of the Executor. So this PR fixes it
properly by introducing a handle type that keeps a reference to some
shared storage. Executor can hand out its handle to `TrySchedule` and
`ScheduleAfter`. A handle contains a raw pointer to the executor and a
mutex for synchronization purposes.

When an Executor gets destructed it will first invalidate the handle by
setting the pointer in the shared storage to nullptr. Since the lifetime
of the shared storage is independent of the Executor's lifetime we won't
have the same problem as before.

This allows us to make all executors have automatic storage duration (no
shared pointer needed anymore). This greatly simplifies its usage.