Move summary logic to PostProcessor (#4708)
Currently, we only calculate the summary if there are more than one
thread involved.  In time range selection, like with inspections, we
don't whether there is more than one thread involved until we go through
all the callstacks selected. This moves the logic for calculating this
to one place.

This also removes tests that tested one thread+summary or multiple
threads+no summary which isn't supposed to happen in practice.