Add SingleThreadExecutor (#4705)
This is adding a new type of `Executor` where tasks will be scheduled on
a single background thread. The Executor is meant for code that
interacts with Qt. The background thread is created and managed by a
`QThread` which makes it interact well with Qt slots that should also
run on the same background QThread.