Remove shared_ptr from Executor (#4706)
After #4670 we don't require Executors anymore to be wrapped in a
`shared_ptr`. So this change removes all those shared_ptrs and replaces
them by values.

This simplifies the code because instances can now be default
constructed.

I had to remove the constness from some member functions because they
need mutable access to the executor in the same class. This used to work
with a shared_ptr.