Remove MainThreadExecutor interface (#4711)
`orbit_base::MainThreadExecutor` used to be an extension to the
`orbit_base::Executor` interface with additional member functions for
waiting on a future.

Currently this functionality is not used anywhere in the code base and
there are other (better) ways to achieve the same. So this commit
removes the interace and also the class `orbit_qt_utils::FutureWatcher`
which implemented the waiting functionality.

In addition this PR renames `orbit_qt_utils::MainThreadExecutorImpl` to
`orbit_qt_utils::MainThreadExecutor`.