Improve module-lookup performance
Previously, we used ModuleIdentifier as two strings as key to
reference modules and store it in maps. Especially creating copies
of the strings when stroring/passing arround the id around, were
quite a big performance hit.

This change introduces a new ModuleIdentifier class, that is just
a number. It can only be created and managed by the new
ModuleIdentifierProvider class.

The change improves the sampling data view creations by 1.25x.

Test: Manually tested on live and saved captures.