Remove second consecutive resizeSections operation
Resizing the sections of an OrbitTreeView is actually very
expensive for large data views (e.g. many samples), as
Qt runs over all the entries and calculates the text
width.

This expensive operation was done always twice in a row
for the sampling report, as the data view panal
already performs the exact same call in the
Initialize method.

Removing this calls gains a 1.2x speedup when generating
the data views.

This change is not (yet) improving the sizing of
the columns itself.

Test: Measure a capture with many samples.