Use orbit_base::Result to implement CanceledOr/NotFoundOr (#4713)
This is an alternative implementation of #4669. I'd like to know what
you think about it.

It is using `Result<T,E>` (what we also use for `ErrorMessageOr<T>`) to
implement `CanceledOr<T> = Result<T,Canceled>` and `NotFoundOr<T> =
Result<T,NotFound>`.

There are some behavioural changes though: `Result` disables its
generic constructor if the list of arguments would form a valid construction
for both the value and the error type. This is also true for the default
construction. That means `CanceledOr/NotFoundOr` can't have the
same default construction behaviour - which is probably a good thing
though.

In older versions of the Outcome library the residual type of a
`outcome::result` needed to be default constructible. That's not the
case anymore for later versions, but the version we use on the CI
still has this constrained.