Optimize rendering of text for timers.
Previously the fitting of the text into the boxes for the timers took
too long. Root cause for this is a fairly complicated computation for
determining the width of a rendered text. This seems to be used inside
the method to elide text as well.
This PR replaces the QFontMetrics::elidedText and
QFontMetrics::horizontalAdvance by a simplified, less accurate method.
We compute a lookup table storing the rendered width of all ascii
characters. The width is computed as the sum of the width of all
characters in a string (plus a heuristic to obtain an upper bound of the
actual width). The method to elide text is is implemented on basis of
this simplified width calculation.
Compare https://github.com/google/orbit/issues/4626.