Add SelectionData for tracking selections
SelectionData contains the underlying data that any of the sampling tabs
may need.  This makes it easier and quicker to swap between the data of
a full capture, a selection, or an inspection.  The inspection isn't
taking full advantage of this data structure yet as it's still branched,
but I hope to change this in a future PR.

Bug: http://b/264437402