Remove faulty CMake dependency
`OrbitSshQtTests` was depending on `Qt::Test`. This target usually
does not exist. The proper name is `Qt5::Test`.

I guess this worked because on some platforms `Qt::Test` is an alias for
`Qt5::Test`.

To fix that I removed the dependency entirely since the tests are not
directly depending on `Qt5::Test` anymore.

Fix: #4734