Fix to QtTextRenderer.
Remove introspection scopes - looks like the one in RenderLayer was
enough to slow down the run when introspection is enabled.

Fix issue with the lookup functions. Previously we used a char as an
index ignoring that this is a signed type. Implicitly casting a negative
char to size_t would result in a huge number. This merely didn't occur
since we only fed printable char into the lookup.