Introduce batch render groups (#4728)
* Add batch render groups

This will replace the layers in all Batchers and TextRenderers in the follow-up commits. The idea is:

* A render group is no longer only identified by its z-value, but by a
freetext name.
* Render groups can store additional metadata per name. They currently
only store bounding box information to restrict the drawing to this,
but this can be easily extended by e.g. positions etc
* Sorting of the group ids (which is the combined name and layer) assures
correct rendering order
* Management of metadata is not part of the batcher, but in a separate
class as it should be shared across multiple batchers and text renderers

Currently these classes are only used in tests, integration will happen
in a follow-up commit.