Make Future::ThenIfSuccess work with arbitrary Results
`Future::ThenIfSuccess` used to only work with a `Future<ErrorMessageOr<T>>`.
This change generalizes that and make arbitrary `Future<Result<T,E>>`
work.

As a preliminary step, we make `ErrorMessageOr<T>` a type alias to
`Result<T, ErrorMessage>`. It used to be a sub class because we
concluded conversion to bool shouldn't be allowed. I moved this
behaviour change into `Result` and made `ErrorMessageOr` a type alias -
similar to `CanceledOr` and `NotFoundOr`.

The rest of the change is rather mechanical.