Add VerticalSizer for track header (#4752)
Add vertical sizer class to resize new track header. The cursor
changes to horizontal arrows when the mouse is in the sizer.
We keep a minimum width equal to the height of the timeline
for symmetry, to keep the triangle toggles visible, and to keep
the track moving feature available.

Also, set right margin to 0, but keep logic just in case, and tweak
distance between timeline and tracks for more stream lined look.