Add BatchRenderGroupId to Batcher and TextRenderer (#4753)
This replaces the layers, previously only defined by z-values,
with BatchRenderGroupId. The Batcher now uses those to
group primitives into draw calls, and applies the requested
clipping area if any is given.

This PR does not yet change any behavior in CaptureViewElement
and there is no change in existing groups. All stencils stay disabled
for now, so this should not have any visual not performance
impact.

Comparison of visuals & rendering times:
Before - After.

The performance numbers have to be taken with a grain of salt,
as they do fluctuate quite a bit. Some performance regression is
likely expected as the hash of BatchRenderGroupId is not as
cheap as the old float-hash.