Rename HasError to HasErrorWithMessage
`HasError` is a GMock matcher that checks whether a result type contains
an error. When given a value matcher as an argument (`HasError(matcher)`),
the matcher is not applied to the actual error type but is expected to
be a string-like type and it will be checked if the given string is
contained in the error message.

This is surprising because that's not how all the other GMock matchers
work (See `testing::Optional` as an example.)

So this change renames the matcher to `HasErrorWithMessage` and adds a
new matcher `HasError` that works as described/expected. The matcher
given as an argument to `HasError` will be applied to the error type
of the result type (`ErrorMessage`, `Canceled`, and `NotFound` are our
most common error types.)

It also updated the implementation of `HasErrorWithMessage`. This now
uses matcher instead of `absl::StrContains` which results in better
error messages.