Make TimeoutOr a Result type (#4754)
Similar to `Canceled` and `NotFound`, `Timeout` now becomes a result
type and `TimeoutOr<T>` an alias for `Result<T, Timeout>`.

This also adds a helper function `WhenValueOrTimeout` which is a wrapper
around `orbit_base::WhenAny`. But `WhenAny` returns a generic `std::variant<T,
Timeout>` so, there is some glue code which converts it into a
`Result,<T, Timeout>`.