Rename orbit_qt_test_utils::WaitFor (#4768)
I want to introduce a function `orbit_qt_utils::WaitFor` which will be
available in the general code base (not just in tests). It has a
different behaviour than the existing `orbit_qt_test_utils::WaitFor`
because it doesn't integrate an automatic timeout. So I want to
rename `orbit_qt_test_utils::WaitFor` first to avoid any confusion
between those two functions.

So this change renames `orbit_qt_test_utils::WaitFor` to
`orbit_qt_test_utils::WaitForWithTimeout`. It also changes the
return type of `WaitForWithTimeout` to `TimeoutOr<T>` which is a generic
result type that we use for timeouts. It replaces a custom type
created only for `WaitForWithTimeout`, so this removes some code
duplication.