Assure that visible tracks are only accessed in the main thread (#4707)
* Assure that visible tracks are only accessed in the main thread

In this PR we are fixing a bug. AddFrameTrack() was executed in the capture_thread but was modifying the sorted_tracks_ array that should be only modified in the main_thread. Since sorting was executed once per second, I don't expect to have a single issue in the past but it's better to fix it.

So, in this PR we are doing a few things:
- Assure all methods access visible_tracks_ and sorted_tracks only in the main_thread_
- Delay the insertion of FrameTracks in a similar way that we delay its deletion.
- Improve our comments to explain which containers are accessed by both threads and which one should be accessed only by main_thread_

* Fixup: main_thread_id