Remove string members from CallTreeView
The function nodes of the CallTreeView always stored three strings,
for module path, build id and function name. Many of them were
even highly redundant.

As we build the complete tree in memory, we not only allocate/deallocate
a big amount of memory on the heap, but als querry function/module
names for all nodes, even if they are not visible.

In general, Qt is very good at caching and only queries item view
items that are visible.

So this change removes the three strings, and instead retrieve
the content on query.

This reduces the time to update the UI in the 5min test capture
from 29sec to 17sec.

Test: Take a capture of Orbit loading the 5min test capture.